add: schemas refactor pages and routes, adjust context to fix login in private routes without authentication

+++ src/pages/employee-list/EmployeeList.jsx
@@ -1,34 +1,48 @@
 import React, { useEffect, useState } from "react";
+import Lottie from 'react-lottie';
 import { Col } from "react-bootstrap";
-import { alertRequest, MESSAGE_DELETE_ERROR, MESSAGE_DELETE_SUCCESS } from "../../utils/index";
+import { alertConfirmResquest } from "../../utils/utils";
+import loadingLottie  from "../../assets/loading-lottie.json"
 import { employeeCollection, getDocument, deleteDocument, document, db } from "../../db/firebase";
+import { showMessageRequest, DELETE_ERROR_MESSAGE, DELETE_SUCCESS_MESSAGE } from "../../utils/index";
 import { ButtonAction, TableEmployee, TitleTableEmployee, ContainerTableEmployee, FigureImage } from "./Employee.List.styled";
 
 export const EmployeeList = () => {
   const [employees, setEmployees] = useState([]);
-  const[isLoading, setIsLoading] = useState(true);
+  const [isLoading, setIsLoading] = useState(true);
 
   useEffect(() => {
-    (async () => {
-      const snapshot = await getDocument(employeeCollection);
-      const employees = snapshot.docs.map((doc) => ({
-        id: doc.id,
-        ...doc.data(),
-      }));
-      setEmployees(employees);
-      setIsLoading(false);
-    })();
+    const fetchEmployees = async () => {
+      try {
+        const snapshot = await getDocument(employeeCollection);
+        const employees = snapshot.docs.map((doc) => ({
+          id: doc.id,
+          ...doc.data(),
+        }));
+        setEmployees(employees);
+        setIsLoading(false);
+      } catch (error) {
+        console.error(error);
+      } finally {
+        setIsLoading(false);
+      }
+    };
+    fetchEmployees();
   }, []);
 
   const handleDeleteEmployee = async (id) => {
+    const confirmed = await alertConfirmResquest();
+    if (!confirmed) {
+      return;
+    }
     try {
       const employeeRef = document(db, "employee", id);
       await deleteDocument(employeeRef);
       const updatedEmployees = employees.filter((employee) => employee.id !== id);
       setEmployees(updatedEmployees);
-      alertRequest(MESSAGE_DELETE_SUCCESS);
+      showMessageRequest(DELETE_SUCCESS_MESSAGE);
     } catch (error) {
-      alertRequest(MESSAGE_DELETE_ERROR);
+      showMessageRequest(DELETE_ERROR_MESSAGE);
       console.error(error);
     }
   };
@@ -37,66 +51,81 @@ export const EmployeeList = () => {
     <section>
       <ContainerTableEmployee fluid>
         <TitleTableEmployee>Funcionários Cadastrados</TitleTableEmployee>
-        <div></div>
-        <TableEmployee responsive striped bordered hover variant="dark">
-          <thead>
-            <tr>
-              <th>#</th>
-              <th>Foto</th>
-              <th>Nome</th>
-              <th>Email</th>
-              <th>CPF</th>
-              <th>Data de Contratação</th>
-              <th>Status</th>
-              <th>Endereço</th>
-              <th>Ações</th>
-            </tr>
-          </thead>
-          <tbody>
-            {employees.map((employee, index) => (
-              <tr key={employee.id}>
-                <td>{index + 1}</td>
-                <td>
-                  {employee.photoUrl && (
-                    <FigureImage
-                      src={employee.photoUrl}
-                      alt="Foto do funcionário"
-                      width={100}
-                      height={100}
-                    />
-                  )}
-                </td>
-                <td>{employee.name}</td>
-                <td>{employee.email}</td>
-                <td>{employee.cpf}</td>
-                <td>{employee.hiringDate}</td>
-                <td>{employee.status ? "Ativo" : "Inativo"}</td>
-                <td>
-                  <div>
-                    {employee.address.street}, {employee.address.number}
-                  </div>
-                  <div>
-                    {employee.address.neighborhood}, {employee.address.city} -{" "}
-                    {employee.address.state}
-                  </div>
-                  <div>CEP: {employee.address.cep}</div>
-                </td>
-                <Col>
-                  <ButtonAction size="sm" variant="outline-primary">
-                    Editar
-                  </ButtonAction>
-                  <ButtonAction
-                    size="sm"
-                    variant="outline-danger"
-                    onClick={() => handleDeleteEmployee(employee.id)}
-                  >
-                    Excluir
-                  </ButtonAction>
-                </Col>
-              </tr>
-            ))}
-          </tbody>
-        </TableEmployee>
+        {isLoading ? (
+          <div className="d-flex justify-content-center align-items-center">
+            <Lottie
+              options={{ animationData: loadingLottie }}
+              height={500}
+              width={500}
+            />
+          </div>
+        ) : (
+          <>
+            {employees.length > 0 ? (
+              <TableEmployee responsive striped bordered hover variant="dark">
+                <thead>
+                  <tr>
+                    <th>#</th>
+                    <th>Foto</th>
+                    <th>Nome</th>
+                    <th>Email</th>
+                    <th>CPF</th>
+                    <th>Data de Contratação</th>
+                    <th>Status</th>
+                    <th>Endereço</th>
+                    <th>Ações</th>
+                  </tr>
+                </thead>
+                <tbody>
+                  {employees.map((employee, index) => (
+                    <tr key={employee.id}>
+                      <td>{index + 1}</td>
+                      <td>
+                        {employee.photoUrl && (
+                          <FigureImage
+                            src={employee.photoUrl}
+                            alt="Foto do funcionário"
+                            width={100}
+                            height={100}
+                          />
+                        )}
+                      </td>
+                      <td>{employee.name}</td>
+                      <td>{employee.email}</td>
+                      <td>{employee.cpf}</td>
+                      <td>{employee.hiringDate}</td>
+                      <td>{employee.status ? "Ativo" : "Inativo"}</td>
+                      <td>
+                        <div>
+                          {employee.address.street}, {employee.address.number}
+                        </div>
+                        <div>
+                          {employee.address.neighborhood},{" "}
+                          {employee.address.city} - {employee.address.state}
+                        </div>
+                        <div>CEP: {employee.address.cep}</div>
+                      </td>
+                      <Col>
+                        <ButtonAction size="sm" variant="outline-primary">
+                          Editar
+                        </ButtonAction>
+                        <ButtonAction
+                          size="sm"
+                          variant="outline-danger"
+                          onClick={() => handleDeleteEmployee(employee.id)}
+                        >
+                          Excluir
+                        </ButtonAction>
+                      </Col>
+                    </tr>
+                  ))}
+                </tbody>
+              </TableEmployee>
+            ) : (
+              <div>Não há funcionários cadastrados</div>
+            )}
+          </>
+        )}
       </ContainerTableEmployee>
     </section>
   );
